% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{coverage_df}
\alias{coverage_df}
\title{Returns a data-frame giving the coverages for all elements of a 
\linkS4class{GenomicEnrichment} object.}
\usage{
coverage_df(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicEnrichment} object.}
}
\value{
A data-frame giving genome-wide and per-element coverages in 
        nucleotides.
}
\description{
Returns a data-frame giving the coverages for all elements of a 
\linkS4class{GenomicEnrichment} object.
}
