% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{coverage_plot}
\alias{coverage_plot}
\title{Plots the coverages for all elements of a 
\linkS4class{GenomicEnrichment} object.}
\usage{
coverage_plot(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicEnrichment} object.}
}
\value{
A plot of genome-wide and per-element coverages in 
        nucleotides.
}
\description{
Plots the coverages for all elements of a 
\linkS4class{GenomicEnrichment} object.
}
