% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{enrichment_df}
\alias{enrichment_df}
\title{Returns a data-frame giving the enrichments for all elements of a 
\linkS4class{GenomicEnrichment} object.}
\usage{
enrichment_df(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicEnrichment} object.}
}
\value{
A data-frame giving per-element log2-enrichments of all partition 
        categories.
}
\description{
Returns a data-frame giving the enrichments for all elements of a 
\linkS4class{GenomicEnrichment} object.
}
