% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{enrichment_plot}
\alias{enrichment_plot}
\title{Plots the enrichments for all elements of a 
\linkS4class{GenomicEnrichment} object.}
\usage{
enrichment_plot(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicEnrichment} object.}
}
\value{
A plot of per-element log2-enrichments of all partition 
        categories.
}
\description{
Plots the enrichments for all elements of a 
\linkS4class{GenomicEnrichment} object.
}
