% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\docType{class}
\name{GenomicEnrichment-class}
\alias{GenomicEnrichment-class}
\title{GenomicEnrichment objects represent the enrichment of a set of 
genomic regions over a genomic partition.}
\description{
GenomicEnrichment objects represent the enrichment of a set of 
genomic regions over a genomic partition.
}
\section{Slots}{

\describe{
\item{\code{regions}}{A GRangesList object representing the regions on which 
enrichment should be computed.}

\item{\code{partition}}{A GRanges object representing a partition of the genome
to be used as background for the enrichment analysis.}

\item{\code{enrichments}}{A list of data-frames representing the enrichment results
for each elements of \code{regions}.}
}}

