% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{GenomicEnrichment}
\alias{GenomicEnrichment}
\title{Performs region enrichment on a set of regions and returns summarized results.}
\usage{
GenomicEnrichment(queries_regions, genome_partition)
}
\arguments{
\item{queries_regions}{A list of regions whose enrichment ratios must be 
calculated.}

\item{genome_partition}{The genome partition indicating which part of the 
genome fall within which category. Each range should have a 'name' 
attribute indicating its category.}
}
\value{
An object of class \linkS4class{GenomicEnrichment}.
}
\description{
Performs region enrichment on a set of regions and returns summarized results.
}
