% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\docType{methods}
\name{length,GenomicEnrichment-method}
\alias{length,GenomicEnrichment-method}
\title{Returns the number of elements of a \linkS4class{GenomicEnrichment} object.}
\usage{
\S4method{length}{GenomicEnrichment}(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicEnrichment} object.}
}
\value{
The number of elements in \code{x}.
}
\description{
Returns the number of elements of a \linkS4class{GenomicEnrichment} object.
}
