% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\docType{methods}
\name{names,GenomicEnrichment-method}
\alias{names,GenomicEnrichment-method}
\title{Returns the names of the elements of a \linkS4class{GenomicEnrichment} object.}
\usage{
\S4method{names}{GenomicEnrichment}(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicEnrichment} object.}
}
\value{
The names of the elements in \code{x}.
}
\description{
Returns the names of the elements of a \linkS4class{GenomicEnrichment} object.
}
