% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\docType{methods}
\name{names<-,GenomicEnrichment,character-method}
\alias{names<-,GenomicEnrichment,character-method}
\title{Set names of the elements of a \linkS4class{GenomicEnrichment} object.}
\usage{
\S4method{names}{GenomicEnrichment,character}(x) <- value
}
\arguments{
\item{x}{The \linkS4class{GenomicEnrichment} object.}

\item{value}{The new names for the elements of the 
\linkS4class{GenomicEnrichment} object.}
}
\description{
Set names of the elements of a \linkS4class{GenomicEnrichment} object.
}
