% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{proportion_df}
\alias{proportion_df}
\title{Returns a data-frame giving the coverage proportions for all elements of a 
\linkS4class{GenomicEnrichment} object.}
\usage{
proportion_df(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicEnrichment} object.}
}
\value{
A data-frame giving genome-wide and per-element coverage proportions 
        in nucleotides.
}
\description{
Returns a data-frame giving the coverage proportions for all elements of a 
\linkS4class{GenomicEnrichment} object.
}
