% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicEnrichment.R
\name{proportion_plot}
\alias{proportion_plot}
\title{Plots the coverage proportions for all elements of a 
\linkS4class{GenomicEnrichment} object.}
\usage{
proportion_plot(x)
}
\arguments{
\item{x}{A \linkS4class{GenomicEnrichment} object.}
}
\value{
A plot giving genome-wide and per-element coverage proportions 
        in nucleotides.
}
\description{
Plots the coverage proportions for all elements of a 
\linkS4class{GenomicEnrichment} object.
}
