% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{region_enrichment}
\alias{region_enrichment}
\title{Calculates enrichment ratios for query regions against a genome wide
partition of the genome.}
\usage{
region_enrichment(query_regions, genome_partition, genome_order = NULL)
}
\arguments{
\item{query_regions}{The regions whose enrichment ratios must be calculated.}

\item{genome_partition}{The genome partition indicating which part of the genome
fall within which category. Each range should have a 'name' attribute
indicating its category.}

\item{genome_order}{An optional ordering of the region types.}
}
\value{
A data-frame containing the enrichment values.
}
\description{
Calculates enrichment ratios for query regions against a genome wide
partition of the genome.
}
