% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicOverlaps.R
\name{regions}
\alias{regions}
\title{Returns the combined regions from a \linkS4class{GenomicOverlap} object.}
\usage{
regions(x)
}
\arguments{
\item{x}{The \linkS4class{GenomicOverlap} object.}
}
\value{
A \code{GRanges} object representing the combined regions.
}
\description{
Returns the combined regions from a \linkS4class{GenomicOverlap} object.
}
