% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_venn}
\alias{plot_venn}
\title{Generates a venn diagram from a \linkS4class{GenomicOverlap} object.}
\usage{
plot_venn(x, title = NULL)
}
\arguments{
\item{x}{A \linkS4class{GenomicOverlap} object.}

\item{title}{A title for the generated venn diagram.}
}
\value{
The grid object representing the venn.diagram.
}
\description{
Generates a venn diagram from a \linkS4class{GenomicOverlap} object.
}
